+++ component_params.py
@@ -1,0 +1,2 @@
+BOARD_830_PTS_PARAMS = {"dimLine": 66, "dimColumn": 22, "sepAlim": [(0, 4), (0, 18.5)], "sepDistribution": [(0, 10.7)]}
+DIP14_PARAMS = {"pinCount": 14, "chipWidth": 2.4, "label": "DIP 14", "type": "DIP14"}
